+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+### Fix
+- keyServer use keyServerCaller from new-js-sdk
+- remove tokend-js-sdk
+
 ## [1.2.0-rc.6] - 2019-02-20
 ### Added
 - Build version displaying